requirements.txt
python-dotenv==1.2.1
requests==2.32.5
sqlalchemy==2.0.45
pandas==2.3.3